refactor: move d.ts files to .ts files
