+++ src/lib/date.ts
@@ -1,6 +1,6 @@
 import { format } from "date-fns";
 import firebase from "firebase/app";
-import LooseObject from "../../types/LooseObject";
+import LooseObject from "../types/LooseObject";
 
 type DateOrTimestamp = Date | firebase.firestore.Timestamp;
