+++ src/lib/downloadResume.tsx
@@ -6,7 +6,7 @@ import formatResumeFilename from "./formatResumeFilename";
 import createDocx from "./createDocx";
 import getAvatarDataUri from "./getAvatarDataUri";
 import PDFTemplate from "../components/PDFTemplate/PDFTemplate";
-import Resume from "../../types/Resume";
+import { Resume } from "../types/Resume";
 
 export default async function downloadResume(resume: Resume, type = "PDF") {
   const { firstName, lastName, avatar: avatarName } = resume.personalia;
