+++ src/types/docxtemplater-image-module-free/index.d.ts
